:white_check_mark: Add test: cannot purge spell card not found

+++ source/YugiAcademy/YuGiOh/cards/card_management_system.py
@@ -1,5 +1,5 @@
 from YuGiOh.cards import SpellCard
-from assertions import SystemRestrictionInfringed
+from assertions import SystemRestrictionInfringed, DataInconsistencyFound
 
 
 class CardManagementSystem:
@@ -12,8 +12,10 @@ class CardManagementSystem:
 
     def store_spell_card(self, spell_card):
         if self.spell_cards_repository.filter(name=spell_card.name):
-            raise SystemRestrictionInfringed(f'There is already a spell card named {spell_card.name}')
+            raise SystemRestrictionInfringed(f'There is already a spell card named {spell_card.name}.')
         spell_card.save()
 
     def purge_spell_card(self, spell_card):
+        if not self.spell_cards_repository.filter(name=spell_card.name):
+            raise DataInconsistencyFound(f'{spell_card} was expected to be found, but it was not.')
         spell_card.delete()
